requirements.txt
streamlit
pandas
numpy
plotly
psycopg2-binary
sqlalchemy
python-dotenv
psycopg-pool
